Добавление книг

+++ app/Services/UploadPhotoService.php
@@ -8,6 +8,9 @@ class UploadPhotoService {
 
     public static function uploadPhoto($folder, $file, $replacedUrl = null)
     {
+        if (!$file)
+            return null;
+
         Storage::disk('public')->delete($replacedUrl);
         return $file->store($folder, 'public');
     }
